Share the tile fixture setup between the layout bench files
